moving more stuff into protobuf utils and addressing other comments
