--- drivebrain_core_base/src/VehicleMsgConversion.cpp
@@ -1,2 +1,0 @@
-#include <VehicleManager.hpp>
-
